+++ examples/dummybehavior.py
@@ -1,0 +1,39 @@
+import getpass
+import time
+import asyncio
+from spade.agent import Agent
+from spade.behaviour import CyclicBehaviour
+
+
+class DummyAgent(Agent):
+    class MyBehav(CyclicBehaviour):
+        async def on_start(self):
+            print("Starting behaviour . . .")
+            self.counter = 0
+
+        async def run(self):
+            print("Counter: {}".format(self.counter))
+            self.counter += 1
+            await asyncio.sleep(1)
+
+    async def setup(self):
+        print("Agent starting . . .")
+        b = self.MyBehav()
+        self.add_behaviour(b)
+
+
+if __name__ == "__main__":
+    jid = input("JID> ")
+    passwd = getpass.getpass()
+
+    dummy = DummyAgent(jid, passwd)
+    future = dummy.start()
+    future.result()
+
+    print("Wait until user interrupts with ctrl+C")
+    try:
+        while True:
+            time.sleep(1)
+    except KeyboardInterrupt:
+        print("Stopping...")
+    dummy.stop()