+++ examples/periodic_example.py
@@ -1,0 +1,82 @@
+import datetime
+import getpass
+import time
+
+from spade import quit_spade
+from spade.agent import Agent
+from spade.behaviour import CyclicBehaviour, PeriodicBehaviour
+from spade.message import Message
+
+
+class PeriodicSenderAgent(Agent):
+    class InformBehav(PeriodicBehaviour):
+        async def run(self):
+            print(f"PeriodicSenderBehaviour running at {datetime.datetime.now().time()}: {self.counter}")
+            msg = Message(to=self.get("receiver_jid"))  # Instantiate the message
+            msg.body = "Hello World"  # Set the message content
+
+            await self.send(msg)
+            print("Message sent!")
+
+            if self.counter == 5:
+                self.kill()
+            self.counter += 1
+
+        async def on_end(self):
+            # stop agent from behaviour
+            await self.agent.stop()
+
+        async def on_start(self):
+            self.counter = 0
+
+    async def setup(self):
+        print(f"PeriodicSenderAgent started at {datetime.datetime.now().time()}")
+        start_at = datetime.datetime.now() + datetime.timedelta(seconds=5)
+        b = self.InformBehav(period=2, start_at=start_at)
+        self.add_behaviour(b)
+
+
+class ReceiverAgent(Agent):
+    class RecvBehav(CyclicBehaviour):
+        async def run(self):
+            print("RecvBehav running")
+            msg = await self.receive(timeout=10)  # wait for a message for 10 seconds
+            if msg:
+                print("Message received with content: {}".format(msg.body))
+            else:
+                print("Did not received any message after 10 seconds")
+                self.kill()
+
+        async def on_end(self):
+            await self.agent.stop()
+
+    async def setup(self):
+        print("ReceiverAgent started")
+        b = self.RecvBehav()
+        self.add_behaviour(b)
+
+
+if __name__ == "__main__":
+    receiver_jid = input("Receiver JID> ")
+    passwd = getpass.getpass()
+    receiveragent = ReceiverAgent(receiver_jid, passwd)
+
+    sender_jid = input("Sender JID> ")
+    passwd = getpass.getpass()
+    senderagent = PeriodicSenderAgent(sender_jid, passwd)
+
+    future = receiveragent.start(auto_register=True)
+    future.result()  # wait for receiver agent to be prepared.
+
+    senderagent.set("receiver_jid", receiver_jid)  # store receiver_jid in the sender knowledge base
+    senderagent.start(auto_register=True)
+
+    while receiveragent.is_alive():
+        try:
+            time.sleep(1)
+        except KeyboardInterrupt:
+            senderagent.stop()
+            receiveragent.stop()
+            break
+    print("Agents finished")
+    quit_spade()