+++ HISTORY.rst
@@ -2,6 +2,40 @@
 History
 =======
 
+3.2.2 (2021-11-25)
+------------------
+
+* Hotfix for the event loop in windows when python is 3.6
+
+3.2.1 (2021-11-16)
+------------------
+
+* Fixed event loop for windows
+
+3.2.0 (2021-07-13)
+------------------
+
+* Added support for Python 3.8 and 3.9
+* Fixed support for Linux, Windows and macOS
+
+3.1.9 (2021-07-08)
+------------------
+
+* Minor fix in docs.
+
+3.1.8 (2021-07-08)
+------------------
+
+* Added examples.
+* Fixed documentation examples.
+* Added Github CI support.
+
+3.1.7 (2021-06-25)
+------------------
+
+* Added hooks for plugins.
+* Minor bug fixings.
+
 3.1.6 (2020-05-22)
 ------------------
 
@@ -120,4 +154,4 @@ History
 3.0.0 (2017-10-06)
 ------------------
 
-* Started writting 3.0 version.
+* Started writing 3.0 version.